# Optimize merge