--- a/lib/fusuma/plugin/events/records/index_record.rb
+++ b/lib/fusuma/plugin/events/records/index_record.rb
@@ -35,7 +35,7 @@ module Fusuma
             raise "position is NOT body: #{self}" unless mergable?
 
             if records.empty?
-              if Config.find_execute_key(index)
+              if Config.instance.find_execute_key(index)
                 @index = index
                 return self
               end